backend/requirements.txt
Flask
Flask-Restful
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-cors
python-dotenv
celery
redis
Flask-mail
Flask-Caching